Service call routine changes
- Changed the service calls to not be so random but to run every 150 ms
+/- a range.
- Changed the button debouncing to store long key presses.  The old code
allowed the button to be held and then after a short time it acted as if
the button was pressed again.  This new routine now registers one click
as just one.
- Changed the screen frame data[0] from 0xC0 to 0x90 to not turn the
"Disc In" light when there's no disc.
NOTE: The button code seems to act weird in colder weather, I've noticed
issues where the back button stops working, and the enter button and up
never work.