PIDs added & control disabling
- Added a page to check for the MIL and if it's on how many codes are
present (will add support later to display the codes).
- Added a page for clearing codes if there are some present
- Fixed the Boost Pressure and Fuel Pressure calculations.  The
multiplication than division was overflowing the int value and causing a
negative value.  Now I just multiply by a double and this solved the
issue.
- Added the ability to disable the paging so we can use the factory info
screen within switching pages on the CBT.  The way it works is by long
pressing the back button.
- Added very basic long press ability to the button code.